requirements.txt
numpy
pandas
matplotlib
seaborn
ipykernel
jinja2
pytest-playwright
pyppeteer
python-dotenv
selenium
beautifulsoup4
webdriver_manager